cl playPage optimization part1
